requirements.txt
gTTS==2.5.4
openai==1.99.9
praw==7.8.1
python-dotenv==1.1.1
requests==2.32.4
pydub==0.25.1
